+++ src/components/Node.js
@@ -2,16 +2,21 @@ import React, { PropTypes, Component } from 'react'
 import ignoreEvent from '../utils/ignoreEvent'
 import xOfPin from '../utils/xOfPin'
 import computeNodeWidth from '../utils/computeNodeWidth'
-import defaultTheme from './theme'
+import theme from './theme'
 
 class Node extends Component {
   getBody () {
     const {
-      bodyHeight,
       fontSize,
-      pinSize,
+      theme,
       text
     } = this.props
+
+    const {
+      pinSize
+    } = theme
+
+    const bodyHeight = this.props.bodyHeight || theme.nodeBodyHeight
 
     // TODO place an id in the div wrapping the body and try to
     // resolve bodyHeight from its content.
@@ -21,7 +26,7 @@ class Node extends Component {
     TODO The following code works and it is ok for custom nodes.
 
     BUT it os not ok for server side rendering cause foreignobject
-       not supported in image context.
+        is not supported in image context.
 
     return (
       <foreignObject
@@ -65,25 +70,32 @@ class Node extends Component {
 
   render () {
     const {
-      bodyHeight,
       dragged,
       draggedLinkId,
-      color,
       fontSize,
       id,
       ins,
       onCreateLink,
       outs,
-      pinSize,
       selected,
       selectNode,
       text,
+      theme,
       updateLink,
       width,
       willDragNode,
       x,
       y
     } = this.props
+
+    const {
+      highlightColor,
+      nodeBarColor,
+      pinColor,
+      pinSize
+    } = theme
+
+    const bodyHeight = this.props.bodyHeight || theme.nodeBodyHeight
 
     const bodyContent = this.getBody()
 
@@ -108,12 +120,12 @@ class Node extends Component {
         <rect
           fillOpacity={0}
           height={bodyHeight + 2 * pinSize}
-          stroke={(selected || dragged) ? color.selected : color.bar}
+          stroke={(selected || dragged) ? highlightColor : nodeBarColor}
           strokeWidth={1}
           width={computedWidth}
         />
         <rect
-          fill={(selected || dragged) ? color.selected : color.bar}
+          fill={(selected || dragged) ? highlightColor : nodeBarColor}
           height={pinSize}
           width={computedWidth}
         />
@@ -140,7 +152,7 @@ class Node extends Component {
           return (
             <rect
               key={i}
-              fill={color.pin}
+              fill={pinColor}
               height={pinSize}
               onMouseDown={ignoreEvent}
               onMouseUp={onMouseUp}
@@ -151,7 +163,7 @@ class Node extends Component {
         })}
         {bodyContent}
         <rect
-          fill={(selected || dragged) ? color.selected : color.bar}
+          fill={(selected || dragged) ? highlightColor : nodeBarColor}
           height={pinSize}
           transform={`translate(0,${pinSize + bodyHeight})`}
           width={computedWidth}
@@ -169,7 +181,7 @@ class Node extends Component {
           return (
             <rect
               key={i}
-              fill={color.pin}
+              fill={pinColor}
               height={pinSize}
               onClick={ignoreEvent}
               onMouseLeave={ignoreEvent}
@@ -185,23 +197,18 @@ class Node extends Component {
 }
 
 Node.propTypes = {
-  bodyHeight: PropTypes.number.isRequired,
+  bodyHeight: PropTypes.number,
   dragged: PropTypes.bool.isRequired,
   draggedLinkId: PropTypes.string,
-  color: PropTypes.shape({
-    bar: PropTypes.string.isRequired,
-    body: PropTypes.string.isRequired,
-    pin: PropTypes.string.isRequired
-  }).isRequired,
   fontSize: PropTypes.number.isRequired,
   id: PropTypes.string,
   ins: PropTypes.array.isRequired,
   outs: PropTypes.array.isRequired,
   onCreateLink: PropTypes.func.isRequired,
-  pinSize: PropTypes.number.isRequired,
   selected: PropTypes.bool.isRequired,
   selectNode: PropTypes.func.isRequired,
   text: PropTypes.string.isRequired,
+  theme: theme.propTypes,
   updateLink: PropTypes.func.isRequired,
   width: PropTypes.number,
   willDragNode: PropTypes.func.isRequired,
@@ -210,25 +217,15 @@ Node.propTypes = {
 }
 
 Node.defaultProps = {
-  bodyHeight: defaultTheme.nodeBodyHeight,
-  dragged: false,
-  color: {
-    bar: 'lightgray',
-    body: 'whitesmoke',
-    border: 'gray',
-    pin: 'darkgray', // Ahahah darkgray is not darker than gray
-    selected: defaultTheme.highlightColor
-    // Actually we have
-    // whitesmoke < lightgray < darkgray < gray
-  },
+  dragged: false, // TODO looks more like a state
   draggedLinkId: null,
   ins: [],
   onCreateLink: Function.prototype,
   outs: [],
-  pinSize: defaultTheme.pinSize,
   selected: false,
   selectNode: Function.prototype,
   text: 'Node',
+  theme: theme.defaultProps,
   updateLink: Function.prototype,
   willDragNode: Function.prototype
 }